requirements.txt
asgiref==3.7.2
boto3==1.34.64
botocore==1.34.64
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
distlib==0.3.7
Django==5.0.3
djangorestframework==3.14.0
filelock==3.12.2
fonttools==4.50.0
idna==3.6
jmespath==1.0.1
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.8.3
nltk==3.8.1
numpy==1.26.4
packaging==24.0
pandas==2.2.1
Pandas3==0.0.1
pillow==10.2.0
platformdirs==3.10.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
regex==2023.12.25
requests==2.31.0
s3transfer==0.10.1
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
sklearn3==0.0.1
sqlparse==0.4.4
textblob==0.18.0.post0
threadpoolctl==3.3.0
tqdm==4.66.2
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.24.3
xgboost==2.0.3
